fix: move pure functions to models
