+++ src/components/domTransformer/emoji.tsx
@@ -1,13 +1,6 @@
 import { Account } from "@/models/account";
+import { extractEmojiName, isEmojiName } from "@/models/emoji";
 import { intersperse } from "@/utils";
-
-export type EmojiNameFragment<N extends string> = `:${N}:`;
-export function isEmojiName(input: string): input is EmojiNameFragment<string> {
-  return input.startsWith(":") && input.endsWith(":");
-}
-export function extractEmojiName<N extends string>(input: EmojiNameFragment<N>): N {
-  return input.slice(1, -1) as N;
-}
 
 export function transformDisplayNameTags(account: Account) {
   const splitted = Object.keys(account.emojiToUrlMap).reduce(
