+++ src/models/emoji.ts
@@ -11,6 +11,13 @@ import { NodeInfoResolver, detectAPIFamily } from "./nodeinfo";
 import Webfinger from "./webfinger";
 
 const EmojiPattern = /:([^:\/]+):/g;
+export type EmojiNameFragment<N extends string> = `:${N}:`;
+export function isEmojiName(input: string): input is EmojiNameFragment<string> {
+  return input.startsWith(":") && input.endsWith(":");
+}
+export function extractEmojiName<N extends string>(input: EmojiNameFragment<N>): N {
+  return input.slice(1, -1) as N;
+}
 
 export function rewriteHtmlTextEmojis(source: string, emojiToUrlMap: Immutable.Map<string, string>): string {
   return emojiToUrlMap.reduce(
